Arrumando estrutura dos repos conforme o readme
