+++ pointtils/src/main/java/com/pointtils/pointtils/src/application/controllers/PointController.java
@@ -13,8 +13,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.pointtils.pointtils.src.application.controllers.dto.PointRequestDTO;
-import com.pointtils.pointtils.src.application.controllers.dto.PointResponseDTO;
+import com.pointtils.pointtils.src.application.dto.PointRequestDTO;
+import com.pointtils.pointtils.src.application.dto.PointResponseDTO;
 import com.pointtils.pointtils.src.application.services.PointService;
 
 import io.swagger.v3.oas.annotations.Operation;
