+++ pointtils/src/main/java/com/pointtils/pointtils/src/application/dto/PointResponseDTO.java
@@ -1,5 +1,6 @@
-package com.pointtils.pointtils.src.application.controllers.dto;
+package com.pointtils.pointtils.src.application.dto;
 
+import java.time.LocalDateTime;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -10,8 +11,10 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class PointRequestDTO {
+public class PointResponseDTO {
+    private Long id;
     private String userId;
     private String description;
-    private String type; // ENTRY, EXIT, LUNCH_START, LUNCH_END
+    private LocalDateTime timestamp;
+    private String type;
 }
